chore: 人流量分析模块历史人流量接口

+++ detect-service/detect-common/src/main/java/com/explore/common/database/FlowHour.java
@@ -13,7 +13,7 @@ import java.time.LocalDate;
  **/
 @Data
 public class FlowHour {
-    private Integer cameraId;
+    private Long cameraId;
 
     private LocalDate date;
